Redesign of ID / Model concept Split between Server/Client

+++ Minimal-J/src/main/java/ch/openech/mj/db/AbstractIndex.java
@@ -27,7 +27,7 @@ public abstract class AbstractIndex<T> implements Index<T> {
 		this.selectQuery = selectQuery();
 	}
 
-	public T lookup(Integer id) {
+	public T lookup(Long id) {
 		if (id != null) {
 			if (table instanceof ImmutableTable) {
 				return ((ImmutableTable<T>) table).read(id);
