+++ Minimal-J/src/main/java/ch/openech/mj/db/ColumnIndex.java
@@ -20,26 +20,26 @@ public class ColumnIndex<T> extends AbstractIndex<T> {
 		this.innerIndex = innerIndex;
 	}
 
-	private List<Integer> findIds(List<Integer> ids) throws SQLException {
+	private List<Long> findIds(List<Long> ids) throws SQLException {
 		PreparedStatement selectStatement = table.getStatement(dbPersistence.getConnection(), selectQuery, false);
-		List<Integer> result = new ArrayList<>(ids.size());
-		for (Integer i : ids) {
+		List<Long> result = new ArrayList<>(ids.size());
+		for (Long i : ids) {
 			helper.setParameter(selectStatement, 1, i, property);
 			result.addAll(executeSelectIds(selectStatement));
 		}
 		return result;
 	}
 	
-	public List<Integer> findIds(Object query) {
+	public List<Long> findIds(Object query) {
 		Connection connection = dbPersistence.getConnection();
 		try {
 			if (innerIndex != null) {
-				List<Integer> queryIds = innerIndex.findIds(query);
+				List<Long> queryIds = innerIndex.findIds(query);
 				return findIds(queryIds);
 			}
 			PreparedStatement selectStatement = table.getStatement(connection, selectQuery, false);
 			helper.setParameter(selectStatement, 1, query, property);
-			List<Integer> result = executeSelectIds(selectStatement);
+			List<Long> result = executeSelectIds(selectStatement);
 			return result;
 		} catch (SQLException x) {
 			String message = "Couldn't use index of column " + column + " of table " + table.getTableName() + " with query " + query;
@@ -48,19 +48,19 @@ public class ColumnIndex<T> extends AbstractIndex<T> {
 	}
 
 	public List<T> findObjects(Object query) {
-		List<Integer> ids = findIds(query);
+		List<Long> ids = findIds(query);
 		List<T> result = new ArrayList<>(ids.size());
-		for (Integer id : ids) {
+		for (Long id : ids) {
 			result.add(lookup(id));
 		}
 		return result;
 	}
 	
-	private List<Integer> executeSelectIds(PreparedStatement preparedStatement) throws SQLException {
-		List<Integer> result = new ArrayList<>();
+	private List<Long> executeSelectIds(PreparedStatement preparedStatement) throws SQLException {
+		List<Long> result = new ArrayList<>();
 		try (ResultSet resultSet = preparedStatement.executeQuery()) {
 			while (resultSet.next()) {
-				result.add(resultSet.getInt(1));
+				result.add(resultSet.getLong(1));
 			}
 		}
 		return result;
